+++ packages/ente-api/src/auth/basic.strategy.ts
@@ -2,6 +2,8 @@ import { PassportStrategy } from "@nestjs/passport";
 import { BasicStrategy as _BasicStrategy } from "passport-http";
 import { AuthService } from "./auth.service";
 import { UnauthorizedException, Inject, Injectable } from "@nestjs/common";
+import { RequestContextUser } from "../helpers/request-context";
+import { Some } from "monet";
 
 @Injectable()
 export class BasicStrategy extends PassportStrategy(_BasicStrategy) {
@@ -9,13 +11,28 @@ export class BasicStrategy extends PassportStrategy(_BasicStrategy) {
     super();
   }
 
-  async validate(username: string, password: string) {
+  async validate(
+    username: string,
+    password: string
+  ): Promise<RequestContextUser> {
     const user = await this.authService.findUserByCredentials(
       username,
       password
     );
-    return user.cata(() => {
-      throw new UnauthorizedException();
-    }, u => u);
+    return user.cata<RequestContextUser>(
+      () => {
+        throw new UnauthorizedException();
+      },
+      user => {
+        return {
+          getDto: async () => Some(user),
+          childrenIds: user.children.map(c => c.id),
+          displayname: user.displayname,
+          id: user.id,
+          role: user.role,
+          username: user.username
+        };
+      }
+    );
   }
 }