+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Fixed: JWT decoding now works on non-ascii chars, so users like tom.schoß work as well
 * Fixed: Removed malicious dependency on `event-stream` ([event-stream/#116](https://github.com/dominictarr/event-stream/issues/116))
+* Changed: Backend now requests requesting user from database only when needed
 
 ## [v0.8.9] - 2018-11-26